Refactor directory for layout components
